rewrite SingleEditView test to tsx
